feat: tui — `n` new-task auto-enters the created task
Creating a task with the full new-task flow now jumps the attached client
straight into the new task's engine pane (ready to type the first prompt)
instead of only landing the Tasks-list cursor on it. Both surfaces — the
dedicated `kobe new-task` tab and the in-pane overlay — mirror the
prompt-first `f` quick-create's jump, with the repo init-prompt firing on
the fresh session; adopt enters the last worktree. The build-session +
switch-client jump is extracted to a shared `task-enter` helper reused by
quick-task, new-task, and the Tasks pane.
