+++ packages/kobe/src/tui/new-task/host.tsx
@@ -11,11 +11,13 @@
  *
  * Unlike the overlay — which resolves a result the Tasks pane then acts
  * on — this page is its own process, so it performs the create / adopt
- * against its own daemon connection and then exits. tmux closes the
- * window and returns to the previous tab; the new task shows up in every
- * Tasks pane via the daemon subscribe. The create/adopt branch mirrors
- * the Tasks pane's `createTask` exactly so the two surfaces stay in
- * lockstep.
+ * against its own daemon connection, then JUMPS the attached client into
+ * the new (or last-adopted) task — building its session so the user lands
+ * in the engine pane ready to type the first prompt — and exits. tmux
+ * closes this window; the client is already on the new task's session. The
+ * new task also shows up in every Tasks pane via the daemon subscribe. The
+ * create/adopt + auto-enter branch mirrors the Tasks pane's `createTask`
+ * (and `quick-task`'s jump) so the surfaces stay in lockstep.
  */
 
 import { connectOrStartDaemon } from "@sma1lboy/kobe-daemon/client/daemon-process"
@@ -23,10 +25,11 @@ import { onMount } from "solid-js"
 import { RemoteOrchestrator } from "../../client/remote-orchestrator.ts"
 import { availableEngineIds } from "../../engine/account-detect.ts"
 import { addSavedRepo, getPersistedString, getSavedRepos, setPersistedString } from "../../state/repos.ts"
-import { DEFAULT_TASK_VENDOR, type VendorId } from "../../types/task.ts"
+import { DEFAULT_TASK_VENDOR, type Task, type VendorId } from "../../types/task.ts"
 import { NewTaskDialog } from "../component/new-task-dialog"
 import { useTheme } from "../context/theme"
 import { bootPaneHost } from "../lib/host-boot"
+import { ensureTaskSession, jumpToTask } from "../lib/task-enter.ts"
 import { useDialog } from "../ui/dialog"
 
 export interface NewTaskHostArgs {
@@ -70,10 +73,13 @@ export function NewTaskPage(props: NewTaskHostArgs & { orchestrator: RemoteOrche
       process.exit(1)
     }
 
+    let entered: Task | undefined
     try {
       if (result.mode === "adopt") {
+        // Adopt one or more existing worktrees; enter the LAST one (mirrors the
+        // Tasks pane's "focus the last adopted" — KOB-256).
         for (const w of result.adopt) {
-          await orch.adoptWorktree({
+          entered = await orch.adoptWorktree({
             repo: result.repo,
             worktreePath: w.worktreePath,
             branch: w.branch,
@@ -81,7 +87,7 @@ export function NewTaskPage(props: NewTaskHostArgs & { orchestrator: RemoteOrche
           })
         }
       } else {
-        await orch.createTask({
+        entered = await orch.createTask({
           repo: result.repo,
           baseRef: result.baseRef,
           vendor: result.vendor,
@@ -90,6 +96,22 @@ export function NewTaskPage(props: NewTaskHostArgs & { orchestrator: RemoteOrche
     } catch (err) {
       console.error("[kobe new-task] task.create/adopt failed:", err)
       process.exit(1)
+    }
+
+    // Auto-enter the new (or last-adopted) task: build its session — letting
+    // the repo's init-prompt fire as the engine's first message — and jump the
+    // attached client into it, landing the user in the engine pane ready to
+    // type. Then exit; the client is already on the new session, so closing
+    // this window doesn't disturb it. Non-fatal: the task is already created,
+    // so a jump failure (e.g. run outside a kobe tmux session) still exits 0 —
+    // the user can enter it from any Tasks pane.
+    if (entered) {
+      try {
+        await ensureTaskSession(orch, entered, result.repo, result.vendor, { includeInitPrompt: true })
+        await jumpToTask(orch, entered, result.repo, result.vendor)
+      } catch (err) {
+        console.error("[kobe new-task] auto-enter failed:", err)
+      }
     }
     process.exit(0)
   }
